FTH-29: refactor BookmarkScreen and related components for improved structure
